requirement.txt
Django==1.11.2
django-crispy-forms==1.6.1
django-taggit==0.22.1
pytz==2017.2
